feature(#544): Add entity oracle
- Caches, blocks at the moment, though can be extended to any other entity type
  in the arena.

- Added all of the oracular controllers (DPO, MDPO at all depths).

- Retooled loop functions to be more modular/easily support the addition of yet
  MORE controllers (which will inevitably happen...). This was a MASSIVE
  refactor.

- Increased usage of forward decls so that controller typenames only need to be
  directly stated outside of function parameter names once.

+semver: feature

+++ include/fordyca/support/tv/block_op_filter.hpp
@@ -105,7 +105,7 @@ class block_op_filter : public rer::client<block_op_filter<T>> {
       case block_op_src::ekNEW_CACHE_DROP:
         return new_cache_drop_filter(controller, cache_prox_dist);
       default:
-        ER_FATAL_SENTINEL("Unhandled penalty type %d", src);
+        ER_FATAL_SENTINEL("Unhandled penalty type %d", static_cast<int>(src));
     } /* switch() */
     return filter_res_t{};
   }
