+++ include/fordyca/support/tv/cache_op_penalty_handler.hpp
@@ -104,7 +104,7 @@ class cache_op_penalty_handler
             timestep,
             original_penalty(),
             penalty,
-            src);
+            static_cast<int>(src));
 
     penalty_list().push_back(
         temporal_penalty<T>(&controller, id, penalty, timestep));
